requirements.txt
fastapi
uvicorn
httpx
aiofiles
python-dotenv
gradio-client
loguru
pydantic
